move document functions in own class
