+++ src/Document.class.php
@@ -28,7 +28,7 @@ final class Document{
 		$this->PATH=PATH."datasets/documents/".$this->ID;
 		$this->URL=URL.$this->ID;
 		$this->DIR=ROOT.$this->PATH."/";
-		$this->TITLE=wdf_document_title($this->ID);
+		$this->TITLE=self::getTitle($this->ID);
 		$this->FILE=$this->DIR."content.md";
 		$this->TIMESTAMP=null;
 		// check if file exist
@@ -89,7 +89,7 @@ final class Document{
 		// add content or if content is null add document title to source code
 		if($content!=false){$source=$content;}else{$source="# ".$this->TITLE."\n";}
 		// search for sub-documents
-		$sub_documents=wdf_document_index($this->ID);
+		$sub_documents=Document::index($this->ID);
 		// check for elements
 		if(count($sub_documents)){
 			// build sub-documents index
@@ -99,7 +99,7 @@ final class Document{
 				// add element list
 				$source.="- [".$sub_element_fe->label."](".PATH.$sub_element_fe->url.")\n";
 				// search for sub-sub-documents
-				$sub_sub_documents=wdf_document_index($sub_element_fe->url);
+				$sub_sub_documents=Document::index($sub_element_fe->url);
 				// cycle all sub-sub-documents
 				foreach($sub_sub_documents as $sub_sub_element_fe){
 					// add element list
@@ -189,6 +189,174 @@ final class Document{
 		}
 		// return
 		return $hierarchy_array;
+	}
+
+
+
+	/**
+	 * Document title
+	 *
+	 * @param string $document Document ID
+	 * @return string Document title
+	 */
+	static function getTitle(string $document):string{
+		$title='';
+		// make path
+		$content_path=DIR."datasets/documents/".$document."/content.md";
+		// load content line by line to find document title if exist
+		if(file_exists($content_path)){
+			$handle=fopen($content_path,"r");
+			while(!feof($handle)){
+				$line=fgets($handle);
+				if(substr($line,0,2)=="# "){
+					$title=trim(substr($line,1));
+					break;
+				}
+			}
+			fclose($handle);
+		}
+		if(!strlen($title)){
+			// make title by path
+			$hierarchy=explode("/",$document);
+			$title=ucwords(str_replace("-"," ",end($hierarchy)));
+		}
+		// return
+		return $title;
+	}
+
+
+
+	/**
+	 * Document list in path
+	 *
+	 * @param ?string $parent Parent Document ID
+	 * @return Document[] array of documents
+	 */
+	static function list(?string $parent=null):array{
+		$documents_array=array();
+		// check parameters
+		if(substr((string)$parent,-1)!="/"){$parent.="/";}
+		if($parent=="/"){$parent=null;}
+		// make directory full path
+		$directory_path=DIR."datasets/documents/".$parent;
+		// check for directory
+		if(is_dir($directory_path)){
+			// scan directory for documents
+			$elements=scandir($directory_path);
+			// cycle all elements
+			foreach($elements as $element_fe){
+				// skip versioning and files
+				if(in_array($element_fe,array(".","..","versions","homepage"))){continue;}
+				if(!is_dir($directory_path.$element_fe)){continue;}
+				// build directory
+				$document=new stdClass();
+				$document->id=$element_fe;
+				$document->path=$parent.$element_fe;
+				$document->url=URL.$document->path;
+				$document->dir=$directory_path.$element_fe;
+				// add element to documents array
+				$documents_array[]=$document;
+			}
+		}
+		// return
+		return $documents_array;
+	}
+
+
+	/**
+	 * Documents Index
+	 *
+	 * @param ?string $parent Parent document ID
+	 * @return Document[] array of Documents
+	 */
+	static function index(?string $parent=null):array{
+		// definitions
+		$index_array=array();
+		$documents_array=array();
+		// get document list
+		$directories_array=Document::list($parent);
+		// build documents array
+		if(count($directories_array)){
+			// cycle all documents
+			foreach($directories_array as $document_fe){
+				// definitions
+				$document_url=$parent."/".$document_fe->id;
+				$document_label=Document::getTitle($document_url);
+				// check document url
+				if(substr($document_url,0,1)=="/"){$document_url=substr($document_url,1);}
+				// add document to documents array
+				$documents_array[$document_url]=$document_label;
+			}
+		}
+		// sort document array by title
+		asort($documents_array);
+		// cycle all documents and build index array
+		foreach($documents_array as $url_fe=>$label_fe){
+			// build index element
+			$element=new stdClass();
+			$element->label=$label_fe;
+			$element->url=$url_fe;
+			// add element to index array
+			$index_array[]=$element;
+		}
+		// return
+		return $index_array;
+	}
+
+	/**
+	 * Documents search in path
+	 *
+	 * @param string $query String for search
+	 * @param ?string $parent Parent Document ID
+	 * @return array of results
+	 */
+	static function search(string $query,?string $parent=null):array{
+		// tree to array
+		function tree_to_array(&$array,$parent=null){
+			foreach(Document::list($parent) as $dir_fe){
+				//wdf_dump($dir);
+				$array[]=$dir_fe->path;
+				tree_to_array($array,$dir_fe->path);
+			}
+		}
+		// definitions
+		$paths_array=array();
+		$matches_array=array();
+		$queries_array=explode(" ",$query);
+		// check for query or return the empty array
+		if(!count($queries_array)){return $matches_array;}
+		// get all documents directories recursively
+		tree_to_array($paths_array,$parent);
+		//wdf_dump($paths_array);
+		// cycle all directories
+		foreach($paths_array as $path_fe){
+			// check if content file exist
+			if(file_exists(DIR."datasets/documents/".$path_fe."/content.md")){
+				// open file handle for read
+				$handle=fopen(DIR."datasets/documents/".$path_fe."/content.md","r");
+				if($handle){
+					while(!feof($handle)){
+						// get line in buffer
+						$buffer=fgets($handle);
+						// define a buffer id
+						$buffer_id=md5($buffer.rand());
+						// cycle all query words
+						foreach($queries_array as $query_fe){
+							// check for query word
+							if(stripos($buffer,$query_fe)!==false){
+								// highlight query word in buffer
+								$buffer=str_ireplace($query_fe,"<mark>".$query_fe."</mark>",$buffer);
+								$matches_array[$path_fe][$buffer_id]=$buffer;
+								// skip current file after 3 matches
+								if(count($matches_array[$path_fe])>2){continue(3);}
+							}
+						}
+					}
+					fclose($handle);
+				}
+			}
+		}
+		return $matches_array;
 	}
 
 }
